Rename gl2x to gl4x
