+++ simulant/renderers/gl4x/generic_renderer.cpp
