+++ simulant/renderers/renderer_config.h
@@ -21,5 +21,5 @@
 #ifdef SIMULANT_GL_VERSION_1X
 #include "gl1x/gl1x_renderer.h"
 #else
-#include "gl2x/generic_renderer.h"
+#include "gl4x/generic_renderer.h"
 #endif
