+++ simulant/utils/gl_error.cpp
@@ -26,7 +26,7 @@
 
 #ifdef SIMULANT_GL_VERSION_4X
     #ifndef __ANDROID__
-        #include "../renderers/gl2x/glad/glad/glad.h"
+        #include "../renderers/gl4x/glad/glad/glad.h"
     #else
         #include <GLES2/gl2.h>
         #include <GLES2/gl2ext.h>
